+++ CHANGELOG.md
@@ -6,9 +6,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.1]
+
+### Added
+
+- Add `noop` Storage that do nothing
+
+### Fixed
+
+- Add protection on global `window` variable
+
 ## [1.0.0]
 
 First version
 
-[Unreleased]: https://github.com/MacFJA/svelte-persistent-store/compare/1.0.0...HEAD
+[Unreleased]: https://github.com/MacFJA/svelte-persistent-store/compare/1.0.1...HEAD
+[1.0.1]: https://github.com/MacFJA/svelte-persistent-store/releases/tag/1.0.1
 [1.0.0]: https://github.com/MacFJA/svelte-persistent-store/releases/tag/1.0.0